requirements.txt
aiohappyeyeballs  2.6.1
aiohttp           3.11.16
aiosignal         1.3.2
annotated-types   0.7.0
anyio             4.9.0
attrs             25.3.0
bcrypt            3.2.0
blinker           1.9.0
certifi           2025.1.31
cffi              1.17.1
click             8.1.8
colorama          0.4.6
deprecation       2.1.0
Flask             3.0.3
Flask-Bcrypt      1.0.1
Flask-Login       0.6.3
frozenlist        1.5.0
gotrue            2.12.0
h11               0.14.0
h2                4.2.0
hpack             4.1.0
httpcore          1.0.8
httpx             0.28.1
hyperframe        6.1.0
idna              3.10
iniconfig         2.1.0
itsdangerous      2.2.0
Jinja2            3.1.6
MarkupSafe        3.0.2
multidict         6.4.3
numpy             2.2.4
packaging         24.2
pandas            2.2.3
pip               25.0
pluggy            1.5.0
postgrest         1.0.1
propcache         0.3.1
pycparser         2.21
pydantic          2.11.3
pydantic_core     2.33.1
PyJWT             2.10.1
pytest            8.3.5
pytest-mock       3.14.0
python-dateutil   2.9.0.post0
python-dotenv     1.0.1
pytz              2025.2
realtime          2.4.2
setuptools        75.8.0
six               1.17.0
sniffio           1.3.1
storage3          0.11.3
StrEnum           0.4.15
supabase          2.15.0
supafunc          0.9.4
typing_extensions 4.13.2
typing-inspection 0.4.0
tzdata            2025.2
websockets        14.2
Werkzeug          3.1.3
wheel             0.45.1
yarl              1.20.0